int to int64 (for osm purposes)

+++ bidirectional_ch_one_to_n_test.go
@@ -12,8 +12,8 @@ func TestOneToManyShortestPath(t *testing.T) {
 	t.Log("Please wait until contraction hierarchy is prepared")
 	g.PrepareContracts()
 	t.Log("TestShortestPath is starting...")
-	u := 106600
-	v := []int{5924, 81611, 69618, 68427, 68490}
+	u := int64(106600)
+	v := []int64{5924, 81611, 69618, 68427, 68490}
 	correctAns, correctPath := []float64{61089.42195558673, 94961.78959757874, 78692.8292369651, 61212.00481622628, 71101.1080090782}, []int{418, 866, 591, 314, 353}
 	ans, path := g.ShortestPathOneToMany(u, v)
 	for i := range path {
@@ -42,8 +42,8 @@ func BenchmarkShortestPathOneToMany(b *testing.B) {
 		n := int(math.Pow(2, k))
 		b.Run(fmt.Sprintf("%s/%d/vertices-%d", "CH shortest path (one to many)", n, len(g.Vertices)), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				u := 106600
-				v := []int{5924, 81611, 69618, 68427, 68490}
+				u := int64(106600)
+				v := []int64{5924, 81611, 69618, 68427, 68490}
 				ans, path := g.ShortestPathOneToMany(u, v)
 				_, _ = ans, path
 			}
@@ -63,8 +63,8 @@ func BenchmarkOldWayShortestPathOneToMany(b *testing.B) {
 		n := int(math.Pow(2, k))
 		b.Run(fmt.Sprintf("%s/%d/vertices-%d", "CH shortest path (one to many)", n, len(g.Vertices)), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				u := 106600
-				v := []int{5924, 81611, 69618, 68427, 68490}
+				u := int64(106600)
+				v := []int64{5924, 81611, 69618, 68427, 68490}
 				for vv := range v {
 					ans, path := g.ShortestPath(u, v[vv])
 					_, _ = ans, path
