+++ bidirectional_relaxation.go
@@ -5,7 +5,7 @@ import (
 )
 
 // relaxEdgesBiForward Edge relaxation in a forward propagation
-func (graph *Graph) relaxEdgesBiForward(vertex *simpleNode, forwQ *forwardPropagationHeap, prev map[int]int, queryDist []float64) {
+func (graph *Graph) relaxEdgesBiForward(vertex *simpleNode, forwQ *forwardPropagationHeap, prev map[int64]int64, queryDist []float64) {
 	vertexList := graph.Vertices[vertex.id].outEdges
 	costList := graph.Vertices[vertex.id].outECost
 	for i := 0; i < len(vertexList); i++ {
@@ -27,7 +27,7 @@ func (graph *Graph) relaxEdgesBiForward(vertex *simpleNode, forwQ *forwardPropag
 }
 
 // relaxEdgesBiForward Edge relaxation in a backward propagation
-func (graph *Graph) relaxEdgesBiBackward(vertex *simpleNode, backwQ *backwardPropagationHeap, prev map[int]int, revDist []float64) {
+func (graph *Graph) relaxEdgesBiBackward(vertex *simpleNode, backwQ *backwardPropagationHeap, prev map[int64]int64, revDist []float64) {
 	vertexList := graph.Vertices[vertex.id].inEdges
 	costList := graph.Vertices[vertex.id].inECost
 	for i := 0; i < len(vertexList); i++ {
